Renamed hashira-deno to hashira-wasm
