--- packages/hashira-deno/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "hashira-deno"
-description = "Deno adapter for hashira"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-hashira = { path = "../hashira" }
-web-sys = { version = "0.3", features = [
-    "Headers",
-    "Request",
-    "Response",
-    "ResponseInit"
-] }
-wasm-bindgen = "0.2"
-wasm-bindgen-futures = "0.4"
-console_error_panic_hook = "0.1"
-js-sys = "0.3.58"
-wasm-streams = "0.3.0"
-futures = "0.3.28"
-serde-wasm-bindgen = "0.5.0"
-
