+++ templates/with-deno/Cargo.toml
@@ -13,7 +13,7 @@ log = "0.4.17"
 hashira = { path = "../../packages/hashira" }
 serde = { version = "1.0", features = ["derive"] }
 wasm-bindgen = "0.2.84"
-hashira-deno = { path = "../../packages/hashira-deno" }
+hashira-wasm = { path = "../../packages/hashira-wasm" }
 web-sys = { version = "0.3", features = [
     "Headers",
     "Request",
@@ -29,6 +29,7 @@ console_error_panic_hook = "0.1"
 js-sys = "0.3.58"
 once_cell = "1.17.1"
 wasm-logger = { version = "0.2.0" }
+serde-wasm-bindgen = "0.5.0"
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3"
